+++ src/library/midi_encoder.cpp
@@ -47,7 +47,7 @@ MidiDataByte encode_note_on(int channel, int note, float velocity)
 {
     MidiDataByte data;
     data[0] = NOTE_ON_PREFIX | static_cast<uint8_t>(channel);
-    data[1] = note;
+    data[1] = static_cast<uint8_t>(note);
     data[2] = static_cast<uint8_t>(std::round(velocity * MAX_VALUE));
     data[3] = 0;
     return data;
@@ -57,7 +57,7 @@ MidiDataByte encode_note_off(int channel, int note, float velocity)
 {
     MidiDataByte data;
     data[0] = NOTE_OFF_PREFIX | static_cast<uint8_t>(channel);
-    data[1] = note;
+    data[1] = static_cast<uint8_t>(note);
     data[2] = static_cast<uint8_t>(std::round(velocity * MAX_VALUE));
     data[3] = 0;
     return data;
@@ -67,7 +67,7 @@ MidiDataByte encode_poly_key_pressure(int channel, int note, float pressure)
 {
     MidiDataByte data;
     data[0] = POLY_PRES_PREFIX | static_cast<uint8_t>(channel);
-    data[1] = note;
+    data[1] = static_cast<uint8_t>(note);
     data[2] = static_cast<uint8_t>(std::round(pressure * MAX_VALUE));
     data[3] = 0;
     return data;
@@ -77,7 +77,7 @@ MidiDataByte encode_control_change(int channel, int controller, float value)
 {
     MidiDataByte data;
     data[0] = CTRL_CHANGE_PREFIX | static_cast<uint8_t>(channel);
-    data[1] = controller;
+    data[1] = static_cast<uint8_t>(controller);
     data[2] = static_cast<uint8_t>(std::round(value * MAX_VALUE));
     data[3] = 0;
     return data;
